feat: refactor to use new code, in progress
