+++ src/components/course/enrollment/EnrollButtonLabel.jsx
@@ -7,25 +7,27 @@ import {
   COURSE_AVAILABILITY_MAP,
   ENROLL_BUTTON_LABEL_COMING_SOON,
   ENROLL_BUTTON_LABEL_NOT_AVAILABLE,
-} from './data/constants';
+} from '../data/constants';
 import {
   isUserEntitledForCourse,
   isCourseSelfPaced,
   hasTimeToComplete,
   isArchived,
-} from './data/utils';
+} from '../data/utils';
 
 const EnrollButtonLabel = ({
   activeCourseRun,
-  availability,
-  courseUuid,
   isCourseStarted,
-  isEnrollable,
   isUserEnrolled,
-  pacingType,
-  start,
   userEntitlements,
 }) => {
+  const {
+    start,
+    availability,
+    isEnrollable,
+    courseUuid,
+    pacingType,
+  } = activeCourseRun;
   // See https://openedx.atlassian.net/wiki/spaces/WS/pages/1045200922/Enroll+button+and+Course+Run+Selector+Logic
   // for more detailed documentation on the enroll button labeling based off course run states.
   const DATE_FORMAT = 'MMM D, YYYY';
@@ -73,15 +75,10 @@ const EnrollButtonLabel = ({
 };
 
 EnrollButtonLabel.propTypes = {
-  isEnrollable: PropTypes.bool.isRequired,
-  courseUuid: PropTypes.string.isRequired,
-  availability: PropTypes.string.isRequired,
-  isCourseStarted: PropTypes.bool.isRequired,
-  userEntitlements: PropTypes.arrayOf(PropTypes.string).isRequired,
   activeCourseRun: PropTypes.shape().isRequired,
+  isCourseStarted: PropTypes.bool.isRequired,
   isUserEnrolled: PropTypes.bool.isRequired,
-  pacingType: PropTypes.string.isRequired,
-  start: PropTypes.string.isRequired,
+  userEntitlements: PropTypes.arrayOf(PropTypes.string).isRequired,
 };
 
 export default EnrollButtonLabel;
